utils.py: Re-sulfle functions and added them to respective files
