+++ test/test_utils.py
@@ -1,6 +1,7 @@
 from shutil import copy
 
 from SCAutolib import utils, LIB_DUMP_USERS
+from SCAutolib.models.user import User
 
 
 def test_load_user(local_user, tmp_path):
@@ -8,5 +9,5 @@ def test_load_user(local_user, tmp_path):
     copy(local_user.dump_file, LIB_DUMP_USERS.joinpath(
         f"{local_user.username}.json"))
 
-    user = utils.load_user(local_user.username)
+    user = User.load(username=local_user.username)
     assert user
